+++ src/js/bootstrap-italia.esm.js
@@ -23,7 +23,7 @@ export { default as InputSearchAutocomplete } from './plugins/input-search-autoc
 export { default as InputPassword } from './plugins/input-password'
 export { default as ProgressDonut } from './plugins/progress-donut'
 export { default as UploadDragDrop } from './plugins/upload-dragdrop'
-export { default as BackToTop } from './plugins/backToTop'
+export { default as BackToTop } from './plugins/back-to-top'
 export { default as HistoryBack } from './plugins/history-back'
 export { default as Forward } from './plugins/forward'
 export { default as Masonry } from './plugins/masonry'
