+++ types/plugins/back-to-top.d.ts
